pyproject.toml
[project]
name = "aacos-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "qdrant-client (>=1.14.3,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "openai (>=1.88.0,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"

